requirements.txt
certifi==2018.11.29
Click==7.0
Flask==1.0.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
numpy==1.15.4
opencv-contrib-python==3.4.4.19
Pillow==5.3.0
pytesseract==0.2.6
Werkzeug==0.14.1
